fixed tests and updated role to current standards
