+++ molecule/default/molecule.yml
@@ -9,7 +9,7 @@ lint:
   name: yamllint
 platforms:
   - name: stretch
-    box: bento/debian-9.0
+    box: bento/debian-9
   - name: xenial
     box: bento/ubuntu-16.04
 provisioner:
