+++ molecule/default/tests/test_package.py
@@ -1,0 +1,13 @@
+import os
+
+import testinfra.utils.ansible_runner
+
+testinfra_hosts = testinfra.utils.ansible_runner.AnsibleRunner(
+    os.environ['MOLECULE_INVENTORY_FILE']).get_hosts('all')
+
+
+def test_package(host):
+    assert host.package('apache2')
+
+def test_executable(host):
+    assert host.file('/usr/sbin/apache2').exists
