+++ molecule/default/tests/test_vhosts_absent.py
@@ -6,9 +6,5 @@ testinfra_hosts = testinfra.utils.ansible_runner.AnsibleRunner(
     os.environ['MOLECULE_INVENTORY_FILE']).get_hosts('all')
 
 
-def test_hosts_file(host):
-    f = host.file('/etc/hosts')
-
-    assert f.exists
-    assert f.user == 'root'
-    assert f.group == 'root'
+def test_absent_vhost(host):
+    assert not host.file('/etc/apache2/sites-enabled/000-default.conf').exists
