--- molecule/install/playbook.yml
@@ -1,20 +1,0 @@
----
-- name: Converge
-  hosts: all
-  gather_facts: False
-  tasks:
-    - name: Install python for Ansible
-      raw: test -e /usr/bin/python || (apt -y update && apt install -y python-minimal)
-      become: True
-      changed_when: False
-
-- name: Converge
-  hosts: all
-  become: yes
-  vars:
-    apache2_enabled: yes
-    apache2_enable_modules:
-      - ssl
-    apache2_enable_letsencrypt: yes
-  roles:
-    - role: role-apache2
