+++ requirements.txt
@@ -1,3 +1,65 @@
-ansible
-molecule
-python-vagrant
+ansible==2.6.4
+ansible-lint==3.4.23
+anyconfig==0.9.7
+arrow==0.12.1
+asn1crypto==0.24.0
+attrs==17.4.0
+backports.functools-lru-cache==1.5
+bcrypt==3.1.4
+binaryornot==0.4.4
+Cerberus==1.2
+certifi==2018.10.15
+cffi==1.11.5
+chardet==3.0.4
+click==6.7
+click-completion==0.3.1
+colorama==0.3.9
+configparser==3.5.0
+cookiecutter==1.6.0
+cryptography==2.2.2
+enum34==1.1.6
+fasteners==0.14.1
+flake8==3.5.0
+funcsigs==1.0.2
+future==0.16.0
+git-url-parse==1.0.2
+idna==2.6
+ipaddress==1.0.19
+Jinja2==2.10
+jinja2-time==0.2.0
+MarkupSafe==1.0
+marshmallow==2.13.5
+mccabe==0.6.1
+molecule==2.18.1
+monotonic==1.4
+more-itertools==4.1.0
+paramiko==2.4.1
+pathspec==0.5.5
+pbr==4.1.0
+pexpect==4.6.0
+pkg-resources==0.0.0
+pluggy==0.6.0
+poyo==0.4.1
+psutil==5.4.6
+ptyprocess==0.5.2
+py==1.5.3
+pyasn1==0.4.2
+pycodestyle==2.3.1
+pycparser==2.18
+pycrypto==2.6.1
+pyflakes==1.5.0
+PyNaCl==1.2.1
+pytest==3.5.0
+python-dateutil==2.7.2
+python-gilt==1.2.1
+python-vagrant==0.5.15
+PyYAML==3.13
+requests==2.19.1
+sh==1.12.14
+six==1.11.0
+tabulate==0.8.2
+testinfra==1.14.1
+tree-format==0.1.2
+urllib3==1.23
+whichcraft==0.4.1
+yamllint==1.11.1
