--- tasks/apache2.yml
@@ -1,20 +1,0 @@
----
-- name: include distro installer
-  include_tasks: '{{ distro_installer }}'
-  with_first_found:
-    - install/{{ ansible_distribution }}.{{ ansible_lsb.codename }}.yml
-    - install/{{ ansible_distribution }}.yml
-    - install/default.yml
-  loop_control:
-    loop_var: distro_installer
-
-- name: include distro configurator
-  include_tasks: "{{ distro_configurator }}"
-  with_first_found:
-    - "configure/{{ ansible_distribution }}.{{ ansible_lsb.codename }}.yml"
-    - "configure/{{ ansible_distribution }}.yml"
-    - "configure/default.yml"
-  loop_control:
-    loop_var: distro_configurator
-  tags:
-    - configure
