--- templates/etc/apache2/sites-available/default-ssl.conf.j2
@@ -1,20 +1,0 @@
-<VirtualHost _default_:443>
-        DocumentRoot /var/www/html/
-        DirectoryIndex index.html
-        # ServerName has to be set to something, otherwise the ServerName of default-ssl will default to $(hostname -f) and it will match for it!
-        ServerName default.local
-        RedirectMatch permanent ^/(.*) {{ apache_default_vhost_redirect_destination }}
-        ErrorLog ${APACHE_LOG_DIR}/default-http-error.log
-        CustomLog ${APACHE_LOG_DIR}/default-http-access.log combined
-        SSLEngine on
-        SSLCertificateFile      /etc/ssl/certs/ssl-cert-snakeoil.pem
-        SSLCertificateKeyFile /etc/ssl/private/ssl-cert-snakeoil.key
-        #SSLCertificateChainFile /etc/apache2/ssl.crt/server-ca.crt
-        #SSLCACertificatePath /etc/ssl/certs/
-        #SSLCACertificateFile /etc/apache2/ssl.crt/ca-bundle.crt
-        #SSLCARevocationPath /etc/apache2/ssl.crl/
-        #SSLCARevocationFile /etc/apache2/ssl.crl/ca-bundle.crl
-        #SSLVerifyClient require
-        #SSLVerifyDepth  10
-        #SSLOptions +FakeBasicAuth +ExportCertData +StrictRequire
-</VirtualHost>
